Move the page button data into ButtonPage.jsx from App.jsx
